api/requirements.txt
fastapi==0.115.8
pymysql==1.1.1
python-dotenv==1.0.1
uvicorn==0.34.0